requirements.txt
Django==2.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0

